setup react-navigation and setup HOME SCREEN
